% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{assign_int}
\alias{assign_int}
\title{Assign subgroup indicators}
\usage{
assign_int(x, beta)
}
\arguments{
\item{x}{Numeric vector to stratify. Values outside [0, 1] are returned as NA.}

\item{beta}{Integer (>= 2): number of subgroups.}
}
\value{
An integer vector of subgroup indicators (1..beta), with NA where x is NA
  or outside [0, 1].
}
\description{
Partitions x into beta strata using:
- S_1:     (1/3, 1]
- S_k:     (1/(k+2), 1/(k+1)] for k = 2, ..., beta-1
- S_beta:  [0, 1/(beta + 1)]
}
