% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_outcome_vr.R
\name{construct_outcome_vr}
\alias{construct_outcome_vr}
\title{Construct an output for matching}
\usage{
construct_outcome_vr(res, dist_list, Z, dataset)
}
\arguments{
\item{res}{A callrelax output.}

\item{dist_list}{A possibly sparse representation of the first distance matrix.}

\item{Z}{A vector of treatment status.}

\item{dataset}{The original dataset.}
}
\value{
a list comprising the following three elements: 
feasibility: 0/1 depend on the feasibility of the matching problem.

data_with_matched_to_ind: a data frame that is the same as the original data frame, 
except that a column called "matched_to", a column called "distance" and a column called 
"matched_to_index" are added to it. The variable "matched_to" records the order of the 
treated unit within the treatment group for each matched control, and assigns NA 
to controls not matched to any treated unit. The variable "matched_to_ind" then converts 
this order into the original observation index. Variable "distance" records the 
control-to-treated distance between one matched control and the corresponding 
treated unit, and assigns NA to controls that are left unmatched and all treated 
units. If matching is not feasible, NULL will be returned.

data_without_unmatched_controls: a data frame that is the same as data_with_matched_to_ind, 
except that the unmatched controls are removed. If matching is not feasible, NULL will be returned.
}
\description{
Construct an output for matching
}
