% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_1_k_finebalance.R
\name{create_list_from_scratch_fm}
\alias{create_list_from_scratch_fm}
\title{create_list_from_scratch_fm is the variant of create_list_from_scratch.
It is designed for 1-to-k matching with fine balance.}
\usage{
create_list_from_scratch_fm(
  Z,
  X,
  fine,
  k,
  alpha = 1,
  method = "maha",
  dist_func = NULL
)
}
\arguments{
\item{Z}{A length-n vector of treatment indicator.}

\item{X}{A n-by-p numerical data.frame of covariates, including the covariate(s) used for fine balance.}

\item{fine}{The column names of the covariate(s) used for fine balance.}

\item{k}{The number of control units matched to each treated unit.}

\item{alpha}{Tuning parameter.}

\item{method}{Method used to compute treated-control distance.}

\item{dist_func}{A function used to calculate distance.
See documentation of "create_list_from_scratch" for more details.}
}
\value{
This function returns a list of three objects: start_n, end_n, and d.
}
\description{
create_list_from_scratch_fm is the variant of create_list_from_scratch.
It is designed for 1-to-k matching with fine balance.
}
