% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_list_from_scratch_overall_vr.R
\name{create_list_from_scratch_overall_vr}
\alias{create_list_from_scratch_overall_vr}
\title{Create a sparse list representation of treatment-to-control distance matrix}
\usage{
create_list_from_scratch_overall_vr(
  Z,
  X,
  fine = NULL,
  p = NULL,
  caliper_low = NULL,
  caliper_high = NULL,
  k = NULL,
  penalty = Inf,
  ad = NULL,
  dist_func = NULL
)
}
\arguments{
\item{Z}{A length-n vector of treatment indicator.}

\item{X}{A n-by-p numerical data.frame of covariates, including the covariate(s) used for fine balance. 
The nominal covariate used for fine balance can be kept as a single factor; there is no need to create dummy variables.}

\item{fine}{The column names of the covariate(s) used for fine balance.}

\item{p}{A length-n vector on which a caliper applies, e.g. a vector of propensity score.}

\item{caliper_low}{Size of caliper_inf.}

\item{caliper_high}{Size of caliper_sup.}

\item{k}{Connect each treated to the nearest k controls.}

\item{penalty}{Penalty for violating the caliper. Set to Inf by default.}

\item{ad}{Adjustment coefficient applied to \eqn{\kappa_{\max}}{kappa_max}.}

\item{dist_func}{A function used to calculate distance.}
}
\value{
This function returns a list of three objects: start_n, end_n, and d.
}
\description{
Create a sparse list representation of treatment-to-control distance matrix
}
