% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_control.R
\name{force_control}
\alias{force_control}
\title{Force including certain controls in the final matched samples.}
\usage{
force_control(dist_list, Z, include)
}
\arguments{
\item{dist_list}{A distance_list object.}

\item{Z}{A length-n vector of treatment indicator.}

\item{include}{A binary vector indicating which controls must be included (length(include) = sum(1-Z).}
}
\value{
This function returns a distance list object with added edges.
}
\description{
This function processes the given distance list by adding certain zero-cost
edges so that the user-specified controls are forced into the final matched
samples. This function is of little interest to most users.
}
