% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{mahapd}
\alias{mahapd}
\title{simple covariate data preprocessing}
\usage{
mahapd(X, fine, var_tol = 0, qr_tol = 1e-10)
}
\arguments{
\item{X}{A n-by-p numerical data frame.}

\item{fine}{Column name of covariates used for fine balance}

\item{var_tol}{A non-negative numeric tolerance for variance filtering.}

\item{qr_tol}{A numeric tolerance for QR decomposition rank determination.}
}
\value{
A data frame containing only the retained covariates 
  (non-constant and linearly independent columns).
}
\description{
simple covariate data preprocessing
}
