% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_divide_conquer.R
\name{matching_fm}
\alias{matching_fm}
\title{Perform 1-to-k matching with fine balance.}
\usage{
matching_fm(Z, X, fine, data, k, method = "maha")
}
\arguments{
\item{Z}{A length-n vector of treatment indicator.}

\item{X}{A n-by-p numerical data.frame of covariates, including the covariate(s) used for fine balance.}

\item{fine}{The column names of the covariate(s) used for fine balance.}

\item{data}{The original dataset used to construct final results.}

\item{k}{The number of control units matched to each treated unit.
(Or the number of treated units matched to each control unit if there are more treated units.)}

\item{method}{Method used to compute treated-control distance.}
}
\value{
This function returns a data frame that is the same as the input "data", 
except that a column called "distance" and a column called "matched_to_index" are added to it. 
Or NULL if such matching is infeasible.
}
\description{
Perform 1-to-k matching with fine balance.
}
\examples{
\dontrun{
data("dt_rhc")
data("dt_rhc_ac")
p <- glm(z ~ . - ninsclasMedicare 
              - `ninsclasMedicare & Medicaid` 
              - `ninsclasNo insurance` 
              - `ninsclasPrivate` 
              - `ninsclasPrivate & Medicare`,
              data = dt_rhc_ac, family = binomial())
dt_rhc_ac$p <-  predict(p, type = "response")
Group_index <- assign_int(dt_rhc_ac$p,beta = 4)
dt_rhc_ac$subgroup <- Group_index

dt_rhc$p <- dt_rhc_ac$p
dt_rhc$subgroup <- dt_rhc_ac$subgroup
dt_rhc_1 <- dt_rhc[dt_rhc$subgroup==1,]
dt_rhc_ac_1 <- dt_rhc_ac[dt_rhc_ac$subgroup == 1,] 
dt_rhc_ac_1_cov <- dt_rhc_ac_1[, setdiff(colnames(dt_rhc_ac_1), c("z","subgroup")), drop = FALSE]

A_matched_1 <- matching_fm(Z = dt_rhc_ac_1$z, X = dt_rhc_ac_1_cov, 
                          fine = c("ninsclasMedicare",
                                   "ninsclasMedicare & Medicaid",
                                   "ninsclasNo insurance",
                                   "ninsclasPrivate",
                                   "ninsclasPrivate & Medicare"),
                          data = dt_rhc_1, k = 1, method = "maha")
}


}
