% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_divide_conquer.R
\name{matching_nf}
\alias{matching_nf}
\title{Perform a pair matching with near fine balance.
Arguments are identical to those of matching_fm except that "lambda", a penalty 
that does a trade-off between two parts of the network, is added and "k" is removed.}
\usage{
matching_nf(Z, X, fine, data, method = "maha", lambda = 1000)
}
\arguments{
\item{Z}{A length-n vector of treatment indicator.}

\item{X}{A n-by-p numerical data.frame of covariates, including the covariate(s) used for fine balance.}

\item{fine}{The column names of the covariate(s) used for fine balance.}

\item{data}{The original dataset used to construct final results.}

\item{method}{Method used to compute treated-control distance.}

\item{lambda}{A penalty that does a trade-off between two parts of the network.}
}
\value{
This function returns a data frame that is the same as the input "data", 
except that a column called "distance" and a column called "matched_to_index" are added to it.
}
\description{
Perform a pair matching with near fine balance.
Arguments are identical to those of matching_fm except that "lambda", a penalty 
that does a trade-off between two parts of the network, is added and "k" is removed.
}
