% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{revert_dist_list_cpp}
\alias{revert_dist_list_cpp}
\title{Revert a treated-to-control distance list.}
\usage{
revert_dist_list_cpp(n_t, n_c, startn, endn, d)
}
\arguments{
\item{n_t}{Number of treated units}

\item{n_c}{Number of control units}

\item{startn}{Vector of starting nodes of edges}

\item{endn}{Vector of ending nodes of edges}

\item{d}{Vector of cost associated with edges}
}
\value{
The function returns a list of three vectors:
        startn, endn, and cost.
}
\description{
Revert a treated-to-control distance list.
}
