% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_network_flow.R
\name{solve_network_flow}
\alias{solve_network_flow}
\title{Solve a network flow problem.}
\usage{
solve_network_flow(net)
}
\arguments{
\item{net}{A list of five vectors: startn, endn, ucap, cost, b.}
}
\value{
If the problem is feasible, function returns a list with the following elements:
crash: an integer, equal to zero if the algorithm ran correctly and equal to 1 if it crashed.
feasible: an integer, equal to zero if the problem is not feasible.
x: a vector equal in length to the number of arcs in argument problem net,
giving in each coordinate the number of units of flow passing across the
corresponding edge in the optimal network flow.
If the problem is not feasible, it returns "Not feasible."
}
\description{
This function solves network flow optimization problems
by calling the RELAX-IV algorithm implemented in FORTRAN
by Dimitri Bertsekas and Paul Tseng, and made available by
Sam Pimentel in the package rcbalance.
}
\details{
This function is of limited interest to users.
}
