% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treated_control_net.R
\name{treated_control_net}
\alias{treated_control_net}
\title{Create a treate-to-control network to be solved via a network flow algorithm.}
\usage{
treated_control_net(n_t, n_c, dist_list, controls = 1)
}
\arguments{
\item{n_t}{Number of treated subjects.}

\item{n_c}{Number of controls.}

\item{dist_list}{A list representation of the distance matrix.}

\item{controls}{Number of controls matched to each treated.}
}
\value{
This function returns a list of five vectors:
startn, endn, ucap, cost, b.
}
\description{
This function takes in a list representation of distance matrix
and create a network structure to be solved.
}
\details{
dist_list is a list consisting of the following three elements:
start_n:  the starting nodes for all edges,
end_n:    the ending nodes for all edges,
d: distance of all treated-control edges.
Function create_dist_list in this package constructs
such a list representation given a user-specified distance function.
}
