% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_1_k_finebalance.R
\name{treated_control_net_fm}
\alias{treated_control_net_fm}
\title{treated_control_net_fm is the variant of treated_control_net.
It is designed for 1-to-k matching with fine balance.}
\usage{
treated_control_net_fm(n_t, n_c, dist_list, k)
}
\arguments{
\item{n_t}{Number of treated units.}

\item{n_c}{Number of control units.}

\item{dist_list}{A list representation of the distance matrix.}

\item{k}{Number of control units matched to each treated unit.}
}
\description{
treated_control_net_fm is the variant of treated_control_net.
It is designed for 1-to-k matching with fine balance.
}
