% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treated_control_net_vr.R
\name{treated_control_net_vr}
\alias{treated_control_net_vr}
\title{Create a treate-to-control network to be solved via a network flow algorithm.}
\usage{
treated_control_net_vr(
  n_t,
  n_c,
  dist_list,
  maxcontrols = NULL,
  mincontrols = 1
)
}
\arguments{
\item{n_t}{Number of treated subjects.}

\item{n_c}{Number of controls.}

\item{dist_list}{A list representation of the distance matrix.}

\item{maxcontrols}{Maximum number of controls matched to each treated.}

\item{mincontrols}{Minimum number of controls matched to each treated.}
}
\value{
This function returns a list of five vectors:
startn, endn, ucap, cost, b.
}
\description{
dist_list is a list consisting of the following three elements:
start_n:  the starting nodes for all edges,
end_n:    the ending nodes for all edges,
d:        distance of all treated-control edges.
}
